Changed string comparisons to use `OrdinalIgnoreCase` rather than `InvariantCultureIgnoreCase`.
Not only is ordinal comparisons faster since it does a byte by byte comparison, but for the case of dealing with HTTP headers, it's also the more correct comparsion to do as of .NET 2.0 (see the [MSDN recommendations for string comparisons](http://msdn.microsoft.com/en-us/library/ms973919.aspx)).

As an example, try running these two lines of code:

    Console.Write(String.Equals("Password", "Paßword", StringComparison.OrdinalIgnoreCase));
    Console.Write(String.Equals("Password", "Paßword", StringComparison.InvariantCultureIgnoreCase));

Notice the first (ordinal) comparison returns `false` and the second (invariant culture) returns `true`.