Implemented `ExecuteAsync` overload as an extension method of `IRestClient`.
The concrete implementation `RestClient` has a nice `ExecuteAsync` method that's simply a wrapper to a call of another method on `IRestClient`. Since this method can be completely implemented in terms of the `IRestClient` interface, it makes sense to make it an extension method of `IRestClient`. The benefit is that I don't need to pass around a `RestClient` in my own API but I can pass around an `IRestClient`. Makes unit testing easier too, and that comes for free! Added bonus! :)