Added Asynchronous UserState
Asynchronous methods were missing the ability to specify a userState
parameter which would be contained within the relevant callback
function's parameters in some way.
All Asynchronous calls now provide an optional userState parameter and
IHttpResponse and IRestResponse types now have a UserState property
which will be populated with the given state object for the request.