Made BaseUrl property virtual
I have a settings object that is the one true source for `BaseUrl`. I'd rather avoid having to observe changes to that object in order to set the `BaseUrl` property on the `RestClient`. Instead, I have a derived `MyRestClient` that sets `BaseUrl` to the value of my settings object. If `BaseUrl` is virtual, I can do this. :)