Upgraded SimpleJson to 0.25
Since SimpleJson.cs is a linked file, only one project should have a
packages.config entry for it, the one with the actual file. The others
will fail trying to update a linked file.

I removed the SimpleJson packages.config entry for all the other
projects. I left it in RestSharp\packages.config