Guard against null values
Since anyone can directly reference DefaultParameters, it's possible for
a null value to be added. Definitely a programming mistake most likely,
but we seem to have run into this somehow. It might have been due to
weird multi-threading issues like those in #368 so I'm not sure if this
is still a problem. I figured it doesn't hurt to have defense in depth.
:)